Added Entity & Redrawing
LETS GOOOOOOOOOOOOOOOOO
