+++ src/Physics/layer.ts
@@ -17,7 +17,7 @@ export class Layer{
 
     private readonly _engine    // Instance of Box2D to access functions
 
-    private _entities: Array<any>   // Need to find some way that i can set a type that works for
+    public entities: Array<any>     // Need to find some way that i can set a type that works for
                                     // Graphics and Sprites, maybe PIXI Container?
     private _ops: PhysOps
 
@@ -44,7 +44,7 @@ export class Layer{
         };
 
         // Array for entities
-        this._entities = new Array();
+        this.entities = new Array();
         this.world = new this._engine.b2d.b2World(this._ops.gravity)
     }
 
@@ -60,16 +60,33 @@ export class Layer{
         const clamped = Math.min(ms, ops.maxTime);
         this.world.Step(clamped/1000, ops.velIterations, ops.posIterations);
         this.world.ClearForces();
+
+        this.redrawEntities();
     }
 
     // Find entity
     // Takes b2d body but yk, types r weird init
-    async findEntity(e): Promise<Point>{
+    findEntity(e): Point{
         const trans = e.getPos();
         return this._engine.coOrdWorldToPixel(trans.x, trans.y);
     }
 
     addEntity(e){
-        this._entities.push(e);
+        this.entities.push(e);
+    }
+
+    redrawEntities(){
+        // Redraw entity if it has sprite
+        // And its sprites location is diff
+        for(const entity of this.entities){         
+            const entityPos = this.findEntity(entity);
+            const spritePos = entity.sprite.position;
+            if(
+                spritePos.x == entityPos.x &&
+                spritePos.y == entityPos.y 
+            ) continue;
+
+            entity.sprite.position.set(entityPos.x, entityPos.y);
+        }
     }
 }
